Add tests that demonstrate issues with useSidebarOrderedReports
